pyproject.toml
[tool.poetry]
name = "drf-api-deploy"
version = "0.1.0"
description = ""
authors = ["Basma Nizar Mohammed Eid <basmanizar@outlook.com>"]

[tool.poetry.dependencies]
python = "~3.8"
django-cors-headers = "^3.5.0"
django-environ = "^0.4.5"
psycopg2-binary = "^2.8.6"
whitenoise = "^5.2.0"
gunicorn = "^20.0.4"
djangorestframework_simplejwt = "^4.4.0"

[tool.poetry.dev-dependencies]
django = "^3.1.1"
black = "^20.8b1"
djangorestframework = "^3.12.1"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
